Refactoring, changed add and remove player to process multiple users at once, some bugs fixed

+++ campaign_management.py
@@ -1,6 +1,6 @@
 import discord
 from enum import Enum
-from typing import Union, List, Tuple, Optional
+from typing import Union, List, Tuple
 
 
 class ChannelType(Enum):
@@ -16,6 +16,7 @@ CAMPAIGN_CHANNELS: Channels = [("campaign-chronicle", ChannelType.TEXT, True, Fa
                                ("campaign-general", ChannelType.TEXT, True, True),
                                ("reactions", ChannelType.TEXT, True, True),
                                ("dm-notes", ChannelType.TEXT, False, False),
+                               ("bot-commands", ChannelType.TEXT, False, False),
                                ("session-voice", ChannelType.VOICE, True, True),
                                ("other-voice", ChannelType.VOICE, True, True)]
 
@@ -94,28 +95,26 @@ async def make_voice_channel(server, name: str,
     await set_role_perms(channel, dungeon_master_role, True, False, False)
 
 
-async def create(message: discord.Message, campaign_name) -> None:
+async def create_campaign(message: discord.Message, campaign_name) -> None:
     """Creates the category and all chat channels for a D&D Campaign.
     The message author is then promoted to the Campaign's Dungeon Master Role."""
-    await message.channel.send(f"Creating {campaign_name} for {message.author.name}.")
-    # ToDo: Fix typing here.
+    await message.channel.send(f"Creating {campaign_name} for {message.author.name}. Please wait until "
+                               f"an error or success message is returned.")
 
-    server: Optional[discord.Guild] = message.guild
+    # ToDo: Fix typing here.
+    server = message.guild
     if server is None:
         await message.channel.send("Error while trying to create the campaign category.")
         return
 
-    await message.channel.send(f"Creating {campaign_name} Category.")
     new_category = await server.create_category(campaign_name)
 
     await new_category.set_permissions(server.default_role, view_channel=False)
     await new_category.set_permissions(server.default_role, read_messages=False)
 
-    await message.channel.send(f"Creating {campaign_name} player and dungeon master roles.")
     player_role = await server.create_role(name=f"{campaign_name} Player", permissions=PLAYER_PERMS)
     dm_role = await server.create_role(name=f"{campaign_name} Dungeon Master", permissions=DM_PERMS)
 
-    await message.channel.send(f"Creating {campaign_name} channels.")
     for name, channel_type, player_read, player_write in CAMPAIGN_CHANNELS:
         if channel_type == ChannelType.TEXT:
             await make_text_channel(server, name, new_category, player_role, dm_role, player_write, player_read)
@@ -126,35 +125,41 @@ async def create(message: discord.Message, campaign_name) -> None:
     await message.channel.send(f"The campaign {campaign_name} was successfully created!")
 
 
-async def delete(message: discord.Message, campaign_name: str) -> None:
+async def delete_campaign(message: discord.Message, campaign_name: str) -> None:
     """Deletes the given campaign category, along with all the channels and roles."""
+    await message.channel.send(f"Attempting to delete {campaign_name}. Please wait until "
+                               f"an error or success message is returned.")
     category = discord.utils.get(message.guild.categories, name=campaign_name)
     if category is None:
         await message.channel.send(f"No campaign by the name of {campaign_name} exists, "
                                    f"did you write the name correctly?")
         return None
 
-    await message.channel.send("Deleting text channels.")
+    player_role = discord.utils.get(message.guild.roles, name=f"{campaign_name} Player")
+    dm_role = discord.utils.get(message.guild.roles, name=f"{campaign_name} Dungeon Master")
+    # ToDo: Decide if this should stop the deletion or simply ignore the deletion of the roles.
+    if player_role is None or dm_role is None:
+        await message.channel.send(f"The {campaign_name} Player role or {campaign_name} Dungeon Master "
+                                   f"role does not exist.")
+        return None
+
     for text_channel in category.text_channels:
         await text_channel.delete()
-
-    await message.channel.send("Deleting voice channels.")
     for voice_channel in category.voice_channels:
         await voice_channel.delete()
 
-    player_role = discord.utils.get(message.guild.roles, name=f"{campaign_name} Player")
-    dm_role = discord.utils.get(message.guild.roles, name=f"{campaign_name} Dungeon Master")
-    # ToDo: Think about adding try ... except to blocks similar to these.
-    await message.channel.send("Deleting the category.")
+    # ToDo: Think about adding try ... except to blocks similar to these. ^ Also deals with that.
     await category.delete()
-    await message.channel.send("Deleting player and dungeon master roles.")
     await player_role.delete()
     await dm_role.delete()
-    await message.channel.send(f"Campaign {campaign_name} deleted successfully!")
+    # Using author.send in case the author decided to delete the channel they were in. Could probably check for it.
+    await message.author.send(f"Campaign {campaign_name} deleted successfully!")
 
 
-async def rename(message: discord.Message, campaign_name: str, new_name: str) -> None:
+async def rename_campaign(message: discord.Message, campaign_name: str, new_name: str) -> None:
     """Renames the given campaign category and its roles."""
+    await message.channel.send(f"Attempting to rename {campaign_name} into {new_name}, please wait until an error "
+                               f"or success message is returned.")
     server = message.guild
 
     if server is None:
@@ -173,8 +178,8 @@ async def rename(message: discord.Message, campaign_name: str, new_name: str) ->
         await message.channel.send(f"A Player or Dungeon Master role for {campaign_name} does not exist.")
         return None
 
-    await campaign_category.edit(name = new_name)
-    await player_role.edit(name = f"{new_name} Player")
-    await dungeon_master_role.edit(name = f"{new_name} Dungeon Master")
+    await campaign_category.edit(name=new_name)
+    await player_role.edit(name=f"{new_name} Player")
+    await dungeon_master_role.edit(name=f"{new_name} Dungeon Master")
 
     await message.channel.send(f"{campaign_name} was successfully renamed to {new_name}.")
